requirements-eda.txt
torch
torchvision
torchaudio
timm
kornia
albumentations
opencv-python
pillow
matplotlib
pandas
pyarrow
scikit-learn
scipy
seaborn
plotly
umap-learn
imagehash
tqdm
pyyaml
rich
torchmetrics
grad-cam
torchinfo
notebook
ipywidgets
streamlit
openpyxl
